# sdks/sdkwork-drive-backend-sdk/sdkwork-drive-backend-sdk-typescript/generated/server-openapi/source-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDKWork Drive Backend API",
    "version": "0.1.0",
    "x-sdkwork-owner": "sdkwork-drive",
    "x-sdkwork-api-authority": "sdkwork-drive.backend"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/backend/v3/api/drive/quotas": {
      "get": {
        "tags": [
          "drive"
        ],
        "operationId": "quotas.summary",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-drive",
        "x-sdkwork-api-authority": "sdkwork-drive.backend"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "traceId",
          "requestId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "QuotaSummary": {
        "type": "object",
        "required": [
          "tenantId",
          "totalBytes",
          "objectCount"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "totalBytes": {
            "type": "integer",
            "format": "int64"
          },
          "objectCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    }
  },
...
}